@@ -4,14 +4,15 @@ version: 0.0.20
 homepage: https://github.com/lejard-h/google_maps_webservice
 
 environment:
-  sdk: '>=1.20.1 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   http: '>=0.13.0 <1.0.0'
   meta: ^1.1.0
 
 dev_dependencies:
-  test: ^1.16.5
-  pedantic: ^1.4.0
-  coverage: ^0.14.1
-  test_coverage: ^0.5.0
+  test: ^1.25.8
+#  pedantic: ^1.11.1
+  build_runner: ^2.4.9
+  json_serializable: ^6.8.0
+  coverage: ^1.9.2
