@@ -1,43 +1,50 @@
 library google_maps_webservice.utils;
 
 import 'dart:async';
-import 'package:meta/meta.dart';
-import 'package:http/http.dart';
 
-const kGMapsUrl = 'https://maps.googleapis.com/maps/api';
+import 'package:http/http.dart';
+import 'package:meta/meta.dart';
+
+final kGMapsUrl = Uri.parse('https://maps.googleapis.com/maps/api');
 
 abstract class GoogleWebService {
   @protected
   final Client _httpClient;
 
   @protected
-  final String _url;
+  late final Uri _url;
 
   @protected
-  final String _apiKey;
+  final String? _apiKey;
 
   @protected
-  final Map<String, dynamic> _apiHeaders;
+  final Map<String, String>? _apiHeaders;
 
-  String get url => _url;
+  Uri get url => _url;
 
   Client get httpClient => _httpClient;
 
-  String get apiKey => _apiKey;
+  String? get apiKey => _apiKey;
 
-  Map<String, dynamic> get apiHeaders => _apiHeaders;
+  Map<String, String>? get apiHeaders => _apiHeaders;
 
   GoogleWebService({
-    String apiKey,
-    String baseUrl,
-    @required String url,
-    Client httpClient,
-    Map<String, dynamic> apiHeaders,
-  })  : assert(url != null),
-        _url = '${baseUrl ?? kGMapsUrl}$url',
-        _httpClient = httpClient ?? Client(),
+    String? apiKey,
+    required String apiPath,
+    String? baseUrl,
+    Client? httpClient,
+    Map<String, String>? apiHeaders,
+  })  : _httpClient = httpClient ?? Client(),
         _apiKey = apiKey,
-        _apiHeaders = apiHeaders;
+        _apiHeaders = apiHeaders {
+    var uri = kGMapsUrl;
+
+    if (baseUrl != null) {
+      uri = Uri.parse(baseUrl);
+    }
+
+    _url = uri.replace(path: '${uri.path}$apiPath');
+  }
 
   @protected
   String buildQuery(Map<String, dynamic> params) {
@@ -57,7 +64,7 @@ abstract class GoogleWebService {
   void dispose() => httpClient.close();
 
   @protected
-  Future<Response> doGet(String url, {Map<String, dynamic> headers}) {
+  Future<Response> doGet(String url, {Map<String, String>? headers}) {
     return httpClient.get(Uri.parse(url), headers: headers);
   }
 
@@ -65,7 +72,7 @@ abstract class GoogleWebService {
   Future<Response> doPost(
     String url,
     String body, {
-    Map<String, dynamic> headers,
+    Map<String, String>? headers,
   }) {
     final postHeaders = {
       'Content-type': 'application/json',
@@ -75,25 +82,20 @@ abstract class GoogleWebService {
   }
 }
 
-abstract class GoogleDateTime {
-  @visibleForTesting
-  @protected
-  DateTime dayTimeToDateTime(int day, String time) {
-    if (time.length < 4) {
-      throw ArgumentError(
-          "'time' is not a valid string. It must be four integers.");
-    }
-
-    final _now = DateTime.now();
-    // Maps is 0-index DO^W
-    final _weekday = _now.weekday - 1;
-    final _mondayOfThisWeek = _now.day - _weekday;
-    final _computedWeekday = _mondayOfThisWeek + day;
-
-    final _hour = int.parse(time.substring(0, 2));
-    final _minute = int.parse(time.substring(2));
-
-    return DateTime.utc(
-        _now.year, _now.month, _computedWeekday, _hour, _minute);
+DateTime dayTimeToDateTime(int day, String time) {
+  if (time.length < 4) {
+    throw ArgumentError(
+        "'time' is not a valid string. It must be four integers.");
   }
+
+  day = day == 0 ? DateTime.sunday : day;
+
+  final now = DateTime.now();
+  final mondayOfThisWeek = now.day - now.weekday;
+  final computedWeekday = mondayOfThisWeek + day;
+
+  final hour = int.parse(time.substring(0, 2));
+  final minute = int.parse(time.substring(2));
+
+  return DateTime.utc(now.year, now.month, computedWeekday, hour, minute);
 }
