@@ -34,12 +34,13 @@ Future<void> main() async {
   // define optional parameter explicit
   res = await geolocation.getGeolocation(considerIp: false);
 
-  print(res.status);
   if (res.isOkay) {
-    print('Latitude: ${res.location.lat}');
-    print('Longitude: ${res.location.lng}');
+    print('Latitude: ${res.location?.lat}');
+    print('Longitude: ${res.location?.lng}');
     print('Accuracy: ${res.accuracy}');
   } else {
-    print(res.errorMessage);
+    print(res.error?.message);
   }
+
+  geolocation.dispose();
 }
