@@ -1,23 +1,24 @@
 library google_maps_webservice.staticmap.example;
 
-import 'package:google_maps_webservice/places.dart';
+import 'dart:io';
+
 import 'package:google_maps_webservice/staticmap.dart';
 
-final apiKey = 'APIKEY';
+final apiKey = Platform.environment['API_KEY'] ?? '';
 
 void main() {
   final staticMap = StaticMap(
     apiKey,
     markers: List.from([
-      Location(23.721160, 90.394435),
-      Location(23.732322, 90.385142),
+      Location(lat: 23.721160, lng: 90.394435),
+      Location(lat: 23.732322, lng: 90.385142),
     ]),
     path: Path(
       enc: 'svh~F`j}uOusC`bD',
       color: 'black',
     ),
-    scale: 'false',
+    scale: false,
   );
 
-  staticMap.getUrl();
+  print(staticMap.getUrl());
 }
