@@ -1,8 +1,14 @@
 import 'dart:async';
-import 'dart:convert';
+
 import 'package:google_maps_webservice/src/core.dart';
 import 'package:google_maps_webservice/src/directions.dart';
 import 'package:test/test.dart';
+
+final _uri = Uri(
+  scheme: 'https',
+  host: 'maps.googleapis.com',
+  path: 'maps/api/directions/json',
+);
 
 Future<void> main() async {
   final apiKey = 'MY_API_KEY';
@@ -16,28 +22,44 @@ Future<void> main() async {
     group('build url', () {
       test('simple with String origin/destination', () {
         expect(
-            directions.buildUrl(
-                origin: 'Paris, France', destination: 'Marseilles, France'),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=${Uri.encodeComponent('Paris, France')}&destination=${Uri.encodeComponent('Marseilles, France')}&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Paris, France',
+            destination: 'Marseilles, France',
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Paris, France',
+            'destination': 'Marseilles, France',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('simple with Location origin/destination', () {
         expect(
-            directions.buildUrl(
-                origin: Location(23.43, 65.1),
-                destination: Location(62.323, 53.1)),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=23.43,65.1&destination=62.323,53.1&key=$apiKey'));
+          directions.buildUrl(
+            origin: Location(lat: 23.43, lng: 65.1),
+            destination: Location(lat: 62.323, lng: 53.1),
+          ),
+          _uri.replace(queryParameters: {
+            'origin': '23.43,65.1',
+            'destination': '62.323,53.1',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('simple with String/Location origin/destination', () {
         expect(
-            directions.buildUrl(
-                origin: Location(23.43, 65.1),
-                destination: 'Marseilles, France'),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=23.43,65.1&destination=${Uri.encodeComponent('Marseilles, France')}&key=$apiKey'));
+          directions.buildUrl(
+            origin: Location(lat: 23.43, lng: 65.1),
+            destination: 'Marseilles, France',
+          ),
+          _uri.replace(queryParameters: {
+            'origin': '23.43,65.1',
+            'destination': 'Marseilles, France',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('simple with bad type for origin/destination', () {
@@ -49,7 +71,8 @@ Future<void> main() async {
         }
 
         try {
-          directions.buildUrl(origin: Location(23.43, 65.1), destination: 10);
+          directions.buildUrl(
+              origin: Location(lat: 23.43, lng: 65.1), destination: 10);
         } catch (e) {
           expect((e as ArgumentError).message,
               equals("'destination' must be a '$String' or a '$Location'"));
@@ -58,236 +81,378 @@ Future<void> main() async {
 
       test('avoid', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                avoids: [RouteType.tolls]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=tolls&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            avoids: [RouteType.tolls],
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'avoid': 'tolls',
+            'key': apiKey,
+          }).toString(),
+        );
+
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                avoids: [RouteType.highways]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=highways&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            avoids: [RouteType.highways],
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'avoid': 'highways',
+            'key': apiKey,
+          }).toString(),
+        );
+
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                avoids: [RouteType.indoor]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=indoor&key=$apiKey'));
-        expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                avoids: [RouteType.ferries]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=ferries&key=$apiKey'));
-        expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                avoids: [RouteType.indoor, RouteType.ferries]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&avoid=indoor|ferries&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            avoids: [RouteType.indoor, RouteType.ferries],
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'avoid': 'indoor|ferries',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('travel_mode', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                travelMode: TravelMode.bicycling),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&mode=bicycling&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            travelMode: TravelMode.bicycling,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'mode': 'BICYCLING',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                travelMode: TravelMode.driving),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&mode=driving&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            travelMode: TravelMode.driving,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'mode': 'DRIVING',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                travelMode: TravelMode.transit),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&mode=transit&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            travelMode: TravelMode.transit,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'mode': 'TRANSIT',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                travelMode: TravelMode.walking),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&mode=walking&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            travelMode: TravelMode.walking,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'mode': 'WALKING',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('departure_time', () {
         var d = 1343641500;
         expect(
-            directions.buildUrl(
-                origin: 'Toronto', destination: 'Montreal', departureTime: d),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=$d&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            departureTime: d,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'departure_time': '$d',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                departureTime: DateTime.fromMillisecondsSinceEpoch(d * 1000)),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=$d&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            departureTime: DateTime.fromMillisecondsSinceEpoch(d * 1000),
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'departure_time': '$d',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('departure_time with now', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                departureTime: 'now'),
-            equals(
-              'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=now&key=$apiKey',
-            ));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            departureTime: 'now',
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'departure_time': 'now',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('arrival_time', () {
         var d = 1343641500;
         expect(
-            directions.buildUrl(
-                origin: 'Toronto', destination: 'Montreal', arrivalTime: d),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&arrival_time=$d&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            arrivalTime: d,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'arrival_time': '$d',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                arrivalTime: DateTime.fromMillisecondsSinceEpoch(d * 1000)),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&arrival_time=$d&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            arrivalTime: DateTime.fromMillisecondsSinceEpoch(d * 1000),
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'arrival_time': '$d',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('units', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto', destination: 'Montreal', units: Unit.metric),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&units=metric&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            units: Unit.metric,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'units': 'metric',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                units: Unit.imperial),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&units=imperial&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            units: Unit.imperial,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'units': 'imperial',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('traffic_model', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                trafficModel: TrafficModel.bestGuess),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&traffic_model=best_guess&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            trafficModel: TrafficModel.bestGuess,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'traffic_model': 'best_guess',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                trafficModel: TrafficModel.pessimistic),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&traffic_model=pessimistic&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            trafficModel: TrafficModel.pessimistic,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'traffic_model': 'pessimistic',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                trafficModel: TrafficModel.optimistic),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&traffic_model=optimistic&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            trafficModel: TrafficModel.optimistic,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'traffic_model': 'optimistic',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('transit_mode', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                transitMode: [TransitMode.rail]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&transit_mode=rail&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            transitMode: [TransitMode.rail],
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'transit_mode': 'rail',
+            'key': apiKey,
+          }).toString(),
+        );
+
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                transitMode: [TransitMode.bus]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&transit_mode=bus&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            transitMode: [TransitMode.bus],
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'transit_mode': 'bus',
+            'key': apiKey,
+          }).toString(),
+        );
+
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                transitMode: [
-                  TransitMode.tram,
-                  TransitMode.train,
-                  TransitMode.subway
-                ]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&transit_mode=tram|train|subway&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            transitMode: [
+              TransitMode.tram,
+              TransitMode.train,
+              TransitMode.subway
+            ],
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'transit_mode': 'tram|train|subway',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('transit_routing_preference', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                transitRoutingPreference:
-                    TransitRoutingPreferences.lessWalking),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&transit_routing_preference=less_walking&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            transitRoutingPreference: TransitRoutingPreferences.lessWalking,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'transit_routing_preference': 'less_walking',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                transitRoutingPreference:
-                    TransitRoutingPreferences.fewerTransfers),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&transit_routing_preference=fewer_transfers&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            transitRoutingPreference: TransitRoutingPreferences.fewerTransfers,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'transit_routing_preference': 'fewer_transfers',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('waypoints', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                waypoints: [
-                  Waypoint.optimize(),
-                  Waypoint.fromAddress('Paris'),
-                  Waypoint.fromLocation(Location(42.2, 21.3)),
-                  Waypoint.fromPlaceId('ChIJ3S-JXmauEmsRUcIaWtf4MzE'),
-                  Waypoint.fromEncodedPolyline('gfo}EtohhU')
-                ]),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&waypoints=optimize:true|Paris|42.2,21.3|place_id:ChIJ3S-JXmauEmsRUcIaWtf4MzE|enc:gfo}EtohhU:&key=$apiKey'));
+          directions
+              .buildUrl(origin: 'Toronto', destination: 'Montreal', waypoints: [
+            Waypoint.optimize(),
+            Waypoint.fromAddress('Paris'),
+            Waypoint.fromLocation(Location(lat: 42.2, lng: 21.3)),
+            Waypoint.fromPlaceId('ChIJ3S-JXmauEmsRUcIaWtf4MzE'),
+            Waypoint.fromEncodedPolyline('gfo}EtohhU')
+          ]),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'waypoints':
+                'optimize:true|Paris|42.2,21.3|place_id:ChIJ3S-JXmauEmsRUcIaWtf4MzE|enc:gfo}EtohhU:',
+            'key': apiKey,
+          }).toString(),
+        );
       });
 
       test('alternatives', () {
         expect(
-            directions.buildUrl(
-                origin: 'Toronto', destination: 'Montreal', alternatives: true),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&alternatives=true&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            alternatives: true,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'alternatives': 'true',
+            'key': apiKey,
+          }).toString(),
+        );
         expect(
-            directions.buildUrl(
-                origin: 'Toronto',
-                destination: 'Montreal',
-                alternatives: false),
-            equals(
-                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&alternatives=false&key=$apiKey'));
+          directions.buildUrl(
+            origin: 'Toronto',
+            destination: 'Montreal',
+            alternatives: false,
+          ),
+          _uri.replace(queryParameters: {
+            'origin': 'Toronto',
+            'destination': 'Montreal',
+            'key': apiKey,
+          }).toString(),
+        );
       });
     });
 
     test('decode response', () {
-      var response = DirectionsResponse.fromJson(json.decode(_responseExample));
+      var response = DirectionsResponse.fromJson(_responseExample);
 
       expect(response.isOkay, isTrue);
       expect(response.routes, hasLength(equals(1)));
@@ -376,92 +541,63 @@ Future<void> main() async {
   });
 }
 
-final _responseExample = '''
-{
-  "status": "OK",
-  "geocoded_waypoints" : [
-     {
-        "geocoder_status" : "OK",
-        "place_id" : "ChIJ7cv00DwsDogRAMDACa2m4K8",
-        "types" : [ "locality", "political" ]
-     },
-     {
-        "geocoder_status" : "OK",
-        "place_id" : "ChIJ69Pk6jdlyIcRDqM1KDY3Fpg",
-        "types" : [ "locality", "political" ]
-     },
-     {
-        "geocoder_status" : "OK",
-        "place_id" : "ChIJgdL4flSKrYcRnTpP0XQSojM",
-        "types" : [ "locality", "political" ]
-     },
-     {
-        "geocoder_status" : "OK",
-        "place_id" : "ChIJE9on3F3HwoAR9AhGJW_fL-I",
-        "types" : [ "locality", "political" ]
-     }
-  ],
-  "routes": [ {
-    "summary": "I-40 W",
-    "legs": [ {
-      "steps": [ {
-        "travel_mode": "DRIVING",
-        "start_location": {
-          "lat": 41.8507300,
-          "lng": -87.6512600
-        },
-        "end_location": {
-          "lat": 41.8525800,
-          "lng": -87.6514100
-        },
-        "polyline": {
-          "points": "a~l~Fjk~uOwHJy@P"
-        },
-        "duration": {
-          "value": 19,
-          "text": "1 min"
-        },
-        "html_instructions": "Head \u003cb\u003enorth\u003c/b\u003e on \u003cb\u003eS Morgan St\u003c/b\u003e toward \u003cb\u003eW Cermak Rd\u003c/b\u003e",
-        "distance": {
-          "value": 207,
-          "text": "0.1 mi"
-        }
-      }],
-      "duration": {
-        "value": 74384,
-        "text": "20 hours 40 mins"
-      },
-      "distance": {
-        "value": 2137146,
-        "text": "1,328 mi"
-      },
-      "start_location": {
-        "lat": 35.4675602,
-        "lng": -97.5164276
-      },
-      "end_location": {
-        "lat": 34.0522342,
-        "lng": -118.2436849
-      },
-      "start_address": "Oklahoma City, OK, USA",
-      "end_address": "Los Angeles, CA, USA"
-    } ],
-    "copyrights": "Map data ©2010 Google, Sanborn",
-    "overview_polyline": {
-      "points": "points"
+final _responseExample = {
+  'status': 'OK',
+  'geocoded_waypoints': [
+    {
+      'geocoder_status': 'OK',
+      'place_id': 'ChIJ7cv00DwsDogRAMDACa2m4K8',
+      'types': ['locality', 'political']
     },
-    "warnings": [ ],
-    "waypoint_order": [ 0, 1 ],
-    "bounds": {
-      "southwest": {
-        "lat": 34.0523600,
-        "lng": -118.2435600
-      },
-      "northeast": {
-        "lat": 41.8781100,
-        "lng": -87.6297900
+    {
+      'geocoder_status': 'OK',
+      'place_id': 'ChIJ69Pk6jdlyIcRDqM1KDY3Fpg',
+      'types': ['locality', 'political']
+    },
+    {
+      'geocoder_status': 'OK',
+      'place_id': 'ChIJgdL4flSKrYcRnTpP0XQSojM',
+      'types': ['locality', 'political']
+    },
+    {
+      'geocoder_status': 'OK',
+      'place_id': 'ChIJE9on3F3HwoAR9AhGJW_fL-I',
+      'types': ['locality', 'political']
+    }
+  ],
+  'routes': [
+    {
+      'summary': 'I-40 W',
+      'legs': [
+        {
+          'steps': [
+            {
+              'travel_mode': 'DRIVING',
+              'start_location': {'lat': 41.8507300, 'lng': -87.6512600},
+              'end_location': {'lat': 41.8525800, 'lng': -87.6514100},
+              'polyline': {'points': 'a~l~Fjk~uOwHJy@P'},
+              'duration': {'value': 19, 'text': '1 min'},
+              'html_instructions':
+                  'Head \u003cb\u003enorth\u003c/b\u003e on \u003cb\u003eS Morgan St\u003c/b\u003e toward \u003cb\u003eW Cermak Rd\u003c/b\u003e',
+              'distance': {'value': 207, 'text': '0.1 mi'}
+            }
+          ],
+          'duration': {'value': 74384, 'text': '20 hours 40 mins'},
+          'distance': {'value': 2137146, 'text': '1,328 mi'},
+          'start_location': {'lat': 35.4675602, 'lng': -97.5164276},
+          'end_location': {'lat': 34.0522342, 'lng': -118.2436849},
+          'start_address': 'Oklahoma City, OK, USA',
+          'end_address': 'Los Angeles, CA, USA'
+        }
+      ],
+      'copyrights': 'Map data ©2010 Google, Sanborn',
+      'overview_polyline': {'points': 'points'},
+      'warnings': [],
+      'waypoint_order': [0, 1],
+      'bounds': {
+        'southwest': {'lat': 34.0523600, 'lng': -118.2435600},
+        'northeast': {'lat': 41.8781100, 'lng': -87.6297900}
       }
     }
-  } ]
-}
-''';
+  ]
+};
