@@ -1,8 +1,13 @@
 import 'dart:async';
-import 'dart:convert';
-import 'package:google_maps_webservice/src/utils.dart';
-import 'package:test/test.dart';
+
 import 'package:google_maps_webservice/geocoding.dart';
+import 'package:test/test.dart';
+
+final _uri = Uri(
+  scheme: 'https',
+  host: 'maps.googleapis.com',
+  path: 'maps/api/geocode/json',
+);
 
 Future<void> main() async {
   final apiKey = 'MY_API_KEY';
@@ -16,62 +21,108 @@ Future<void> main() async {
     group('build url', () {
       test('address', () {
         expect(
-            geocoding.buildUrl(
-                address: '1600 Amphitheatre Parkway, Mountain View, CA'),
-            equals(
-                '$kGMapsUrl/geocode/json?address=1600%20Amphitheatre%20Parkway%2C%20Mountain%20View%2C%20CA&key=$apiKey'));
+          geocoding.buildUrl(
+            address: '1600 Amphitheatre Parkway, Mountain View, CA',
+          ),
+          _uri.replace(
+            queryParameters: {
+              'address': '1600 Amphitheatre Parkway, Mountain View, CA',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
 
       test('address with bound', () {
         expect(
-            geocoding.buildUrl(
-                address: 'Winnetka',
-                bounds: Bounds(Location(34.172684, -118.604794),
-                    Location(34.236144, -118.500938))),
-            equals(
-                '$kGMapsUrl/geocode/json?address=Winnetka&bounds=34.172684,-118.604794|34.236144,-118.500938&key=$apiKey'));
+          geocoding.buildUrl(
+            address: 'Winnetka',
+            bounds: Bounds(
+              northeast: Location(lat: 34.172684, lng: -118.604794),
+              southwest: Location(lat: 34.236144, lng: -118.500938),
+            ),
+          ),
+          _uri.replace(
+            queryParameters: {
+              'address': 'Winnetka',
+              'bounds': '34.172684,-118.604794|34.236144,-118.500938',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
 
       test('address with language', () {
         expect(
-            geocoding.buildUrl(address: 'Paris', language: 'fr'),
-            equals(
-                '$kGMapsUrl/geocode/json?address=Paris&language=fr&key=$apiKey'));
+          geocoding.buildUrl(address: 'Paris', language: 'fr'),
+          _uri.replace(
+            queryParameters: {
+              'address': 'Paris',
+              'language': 'fr',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
 
       test('address with region', () {
         expect(
-            geocoding.buildUrl(address: 'Toledo', region: 'es'),
-            equals(
-                '$kGMapsUrl/geocode/json?address=Toledo&region=es&key=$apiKey'));
+          geocoding.buildUrl(address: 'Toledo', region: 'es'),
+          _uri.replace(
+            queryParameters: {
+              'address': 'Toledo',
+              'region': 'es',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
 
       test('address with components', () {
         expect(
-            geocoding.buildUrl(address: 'Spain', components: [
-              Component(Component.administrativeArea, 'Toledo')
-            ]),
-            equals(
-                '$kGMapsUrl/geocode/json?address=Spain&components=administrative_area:Toledo&key=$apiKey'));
+          geocoding.buildUrl(
+            address: 'Spain',
+            components: [Component(Component.administrativeArea, 'Toledo')],
+          ),
+          _uri.replace(
+            queryParameters: {
+              'address': 'Spain',
+              'components': 'administrative_area:Toledo',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
 
       test('location', () {
         expect(
-            geocoding.buildUrl(location: Location(34.172684, -118.604794)),
-            equals(
-                'https://maps.googleapis.com/maps/api/geocode/json?latlng=34.172684,-118.604794&key=$apiKey'));
+          geocoding.buildUrl(
+            location: Location(lat: 34.172684, lng: -118.604794),
+          ),
+          _uri.replace(
+            queryParameters: {
+              'latlng': '34.172684,-118.604794',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
 
       test('place_id', () {
         expect(
-            geocoding.buildUrl(placeId: 'f2hf1pn1rjr1'),
-            equals(
-                'https://maps.googleapis.com/maps/api/geocode/json?place_id=f2hf1pn1rjr1&key=$apiKey'));
+          geocoding.buildUrl(placeId: 'f2hf1pn1rjr1'),
+          _uri.replace(
+            queryParameters: {
+              'place_id': 'f2hf1pn1rjr1',
+              'key': apiKey,
+            },
+          ).toString(),
+        );
       });
     });
 
     test('decode response', () {
-      var response = GeocodingResponse.fromJson(json.decode(_responseExample));
+      var response = GeocodingResponse.fromJson(_responseExample);
 
       expect(response.isOkay, isTrue);
       expect(response.results, hasLength(equals(1)));
@@ -88,13 +139,13 @@ Future<void> main() async {
       expect(
           response.results.first.geometry.location.lng, equals(-122.0842499));
       expect(response.results.first.geometry.locationType, equals('ROOFTOP'));
-      expect(response.results.first.geometry.viewport.northeast.lat,
+      expect(response.results.first.geometry.viewport?.northeast.lat,
           equals(37.4238253802915));
-      expect(response.results.first.geometry.viewport.northeast.lng,
+      expect(response.results.first.geometry.viewport?.northeast.lng,
           equals(-122.0829009197085));
-      expect(response.results.first.geometry.viewport.southwest.lat,
+      expect(response.results.first.geometry.viewport?.southwest.lat,
           equals(37.4211274197085));
-      expect(response.results.first.geometry.viewport.southwest.lng,
+      expect(response.results.first.geometry.viewport?.southwest.lng,
           equals(-122.0855988802915));
       expect(response.results.first.placeId,
           equals('ChIJ2eUgeAK6j4ARbn5u_wAGqWA'));
@@ -103,69 +154,59 @@ Future<void> main() async {
   });
 }
 
-final _responseExample = '''
-{
-   "results" : [
-      {
-         "address_components" : [
-            {
-               "long_name" : "1600",
-               "short_name" : "1600",
-               "types" : [ "street_number" ]
-            },
-            {
-               "long_name" : "Amphitheatre Pkwy",
-               "short_name" : "Amphitheatre Pkwy",
-               "types" : [ "route" ]
-            },
-            {
-               "long_name" : "Mountain View",
-               "short_name" : "Mountain View",
-               "types" : [ "locality", "political" ]
-            },
-            {
-               "long_name" : "Santa Clara County",
-               "short_name" : "Santa Clara County",
-               "types" : [ "administrative_area_level_2", "political" ]
-            },
-            {
-               "long_name" : "California",
-               "short_name" : "CA",
-               "types" : [ "administrative_area_level_1", "political" ]
-            },
-            {
-               "long_name" : "United States",
-               "short_name" : "US",
-               "types" : [ "country", "political" ]
-            },
-            {
-               "long_name" : "94043",
-               "short_name" : "94043",
-               "types" : [ "postal_code" ]
-            }
-         ],
-         "formatted_address" : "1600 Amphitheatre Parkway, Mountain View, CA 94043, USA",
-         "geometry" : {
-            "location" : {
-               "lat" : 37.4224764,
-               "lng" : -122.0842499
-            },
-            "location_type" : "ROOFTOP",
-            "viewport" : {
-               "northeast" : {
-                  "lat" : 37.4238253802915,
-                  "lng" : -122.0829009197085
-               },
-               "southwest" : {
-                  "lat" : 37.4211274197085,
-                  "lng" : -122.0855988802915
-               }
-            }
-         },
-         "place_id" : "ChIJ2eUgeAK6j4ARbn5u_wAGqWA",
-         "types" : [ "street_address" ]
-      }
-   ],
-   "status" : "OK"
-}
-''';
+final _responseExample = {
+  'results': [
+    {
+      'address_components': [
+        {
+          'long_name': '1600',
+          'short_name': '1600',
+          'types': ['street_number']
+        },
+        {
+          'long_name': 'Amphitheatre Pkwy',
+          'short_name': 'Amphitheatre Pkwy',
+          'types': ['route']
+        },
+        {
+          'long_name': 'Mountain View',
+          'short_name': 'Mountain View',
+          'types': ['locality', 'political']
+        },
+        {
+          'long_name': 'Santa Clara County',
+          'short_name': 'Santa Clara County',
+          'types': ['administrative_area_level_2', 'political']
+        },
+        {
+          'long_name': 'California',
+          'short_name': 'CA',
+          'types': ['administrative_area_level_1', 'political']
+        },
+        {
+          'long_name': 'United States',
+          'short_name': 'US',
+          'types': ['country', 'political']
+        },
+        {
+          'long_name': '94043',
+          'short_name': '94043',
+          'types': ['postal_code']
+        }
+      ],
+      'formatted_address':
+          '1600 Amphitheatre Parkway, Mountain View, CA 94043, USA',
+      'geometry': {
+        'location': {'lat': 37.4224764, 'lng': -122.0842499},
+        'location_type': 'ROOFTOP',
+        'viewport': {
+          'northeast': {'lat': 37.4238253802915, 'lng': -122.0829009197085},
+          'southwest': {'lat': 37.4211274197085, 'lng': -122.0855988802915}
+        }
+      },
+      'place_id': 'ChIJ2eUgeAK6j4ARbn5u_wAGqWA',
+      'types': ['street_address']
+    }
+  ],
+  'status': 'OK'
+};
