@@ -2,29 +2,31 @@ import 'dart:async';
 import 'package:test/test.dart';
 import 'package:google_maps_webservice/src/utils.dart';
 
-class DateTimeInstance extends GoogleDateTime {}
-
 Future<void> main() async {
   group('Google Maps Utils', () {
-    group('abstract class GoogleDateTime', () {
-      DateTimeInstance utility;
+    group('dayTimeToDateTime', () {
+      test('basic', () {
+        final sundayAt1130 = dayTimeToDateTime(0, '2330');
 
-      setUpAll(() {
-        utility = DateTimeInstance();
+        expect(sundayAt1130.weekday, equals(DateTime.sunday));
+        expect(sundayAt1130.hour, equals(23));
+        expect(sundayAt1130.minute, equals(30));
+
+        final mondayAt1130 = dayTimeToDateTime(1, '2330');
+
+        expect(mondayAt1130.weekday, equals(DateTime.monday));
+        expect(mondayAt1130.hour, equals(23));
+        expect(mondayAt1130.minute, equals(30));
+
+        final saturdayAt1130 = dayTimeToDateTime(6, '2330');
+
+        expect(saturdayAt1130.weekday, equals(DateTime.saturday));
+        expect(saturdayAt1130.hour, equals(23));
+        expect(saturdayAt1130.minute, equals(30));
       });
 
-      group('dayTimeToDateTime', () {
-        test('basic', () {
-          final mondayAt1130 = utility.dayTimeToDateTime(0, '2330');
-
-          expect(mondayAt1130.weekday, equals(1));
-          expect(mondayAt1130.hour, equals(23));
-          expect(mondayAt1130.minute, equals(30));
-        });
-        test('throws for an invalid time argument', () {
-          expect(
-              () => utility.dayTimeToDateTime(0, '230'), throwsArgumentError);
-        });
+      test('throws for an invalid time argument', () {
+        expect(() => dayTimeToDateTime(0, '230'), throwsArgumentError);
       });
     });
   });
